backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.4.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
anthropic==0.34.2
twilio==9.3.3
sendgrid==6.11.0
httpx==0.27.2
apscheduler==3.10.4
dateparser==1.2.0
python-multipart==0.0.9
python-dotenv==1.0.1
pyjwt==2.9.0
bcrypt==4.2.0
